Show simple MSMQ-to-ASB routing via router
