+++ samples/azure-service-bus-netstandard/azure-service-bus-msmq-bridge/ASBS_2/AsbEndpoint/MyCommandHandler.cs
@@ -10,6 +10,6 @@ public class MyCommandHandler :
     public Task Handle(MyCommand message, IMessageHandlerContext context)
     {
         log.Info($"Received MyCommand: {message.Property}");
-        return context.Publish<MyEvent>(@event => { @event.Property = "event from MSMQ endpoint"; });
+        return Task.CompletedTask;
     }
 }
