+++ samples/azure-service-bus-netstandard/azure-service-bus-msmq-bridge/ASBS_2/AsbEndpoint/Program.cs
@@ -7,8 +7,7 @@ class Program
 {
     static async Task Main()
     {
-        const string letters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-        var random = new Random();
+        
 
         Console.Title = "Samples.Azure.ServiceBus.AsbEndpoint";
         var endpointConfiguration = new EndpointConfiguration("Samples.Azure.ServiceBus.AsbEndpoint");
@@ -32,36 +31,11 @@ class Program
 
         #endregion
 
-        #region route-command-via-bridge
-
-        bridge.RouteToEndpoint(typeof(MyCommand), "Samples.Azure.ServiceBus.MsmqEndpoint");
-
-        #endregion
-
-        #region subscribe-to-event-via-bridge
-
-        bridge.RegisterPublisher(typeof(MyEvent), "Samples.Azure.ServiceBus.MsmqEndpoint");
-
-        #endregion
-
         var endpointInstance = await Endpoint.Start(endpointConfiguration)
             .ConfigureAwait(false);
 
-        Console.WriteLine("Press Enter to send a command");
-        Console.WriteLine("Press any other key to exit");
-
-        while (true)
-        {
-            var key = Console.ReadKey().Key;
-            if (key != ConsoleKey.Enter)
-            {
-                break;
-            }
-
-            var prop = new string(Enumerable.Range(0, 3).Select(i => letters[random.Next(letters.Length)]).ToArray());
-            await endpointInstance.Send(new MyCommand { Property = prop }).ConfigureAwait(false);
-            Console.WriteLine($"\nCommand with value '{prop}' sent");
-        }
+        Console.WriteLine("Press any key to exit");
+        Console.ReadKey();
 
         await endpointInstance.Stop()
             .ConfigureAwait(false);
