--- samples/azure-service-bus-netstandard/azure-service-bus-msmq-bridge/ASBS_2/MsmqEndpoint/OtherEventHandler.cs
@@ -1,14 +1,0 @@
-﻿using System.Threading.Tasks;
-using NServiceBus;
-using NServiceBus.Logging;
-
-public class OtherEventHandler : IHandleMessages<OtherEvent>
-{
-    static ILog log = LogManager.GetLogger<OtherEventHandler>();
-
-    public Task Handle(OtherEvent message, IMessageHandlerContext context)
-    {
-        log.Info($"Received OtherEvent: {message.Property}");
-        return Task.CompletedTask;
-    }
-}
