--- samples/azure-service-bus-netstandard/azure-service-bus-msmq-bridge/ASBS_2/Shared/OtherEvent.cs
@@ -1,7 +1,0 @@
-﻿using NServiceBus;
-
-public interface OtherEvent :
-    IEvent
-{
-    string Property { get; set; }
-}
